+++ trunk/BizHawk.Emulation/Computers/Commodore64/Cartridge/Mapper0012.cs
@@ -8,21 +8,21 @@ namespace BizHawk.Emulation.Computers.Commodore64.Cartridge
 		private byte[] bankMain;
 		private byte[][] bankHigh;
 		private byte[] bankHighSelected;
-		private uint bankIndex;
+		private int bankIndex;
 		private byte[] dummyBank;
 
 		// Zaxxon and Super Zaxxon cartridges
 		// - read to 8xxx selects bank 0 in A000-BFFF
 		// - read to 9xxx selects bank 1 in A000-BFFF
 
-		public Mapper0012(List<uint> newAddresses, List<uint> newBanks, List<byte[]> newData)
+		public Mapper0012(List<int> newAddresses, List<int> newBanks, List<byte[]> newData)
 		{
 			bankMain = new byte[0x2000];
 			bankHigh = new byte[2][];
 			dummyBank = new byte[0x2000];
 
 			// create dummy bank just in case
-			for (uint i = 0; i < 0x2000; i++)
+			for (int i = 0; i < 0x2000; i++)
 				dummyBank[i] = 0xFF;
 
 			bankHigh[0] = dummyBank;
@@ -56,14 +56,14 @@ namespace BizHawk.Emulation.Computers.Commodore64.Cartridge
 			return bankHighSelected[addr];
 		}
 
-		public override byte Read8000(ushort addr)
+		public override byte Read8000(int addr)
 		{
-			bankIndex = (addr & (uint)0x1000) >> 12;
+			bankIndex = (addr & 0x1000) >> 12;
 			bankHighSelected = bankHigh[bankIndex];
 			return bankMain[addr];
 		}
 
-		public override byte ReadA000(ushort addr)
+		public override byte ReadA000(int addr)
 		{
 			return bankHighSelected[addr];
 		}
